requirements.txt
bittensor
torch
web3
streamlit
python-dotenv
uniswap-python
redis
aioredis
loguru